+++ cmd/web/core/application.go
@@ -2,8 +2,13 @@ package core
 
 import (
 	"log/slog"
+
+	"newser.app/internal/repository"
+	"newser.app/internal/service"
 )
 
 type App struct {
-	Logger *slog.Logger
+	Logger      *slog.Logger
+	FeedRepo    *repository.NewsfeedSqliteRepo
+	FeedService *service.API
 }
